+++ google/cloud/google_cloud_cpp_common.bzl
@@ -34,6 +34,7 @@ google_cloud_cpp_common_hdrs = [
     "internal/api_client_header.h",
     "internal/attributes.h",
     "internal/backoff_policy.h",
+    "internal/base64_transforms.h",
     "internal/big_endian.h",
     "internal/build_info.h",
     "internal/compiler_info.h",
@@ -85,6 +86,7 @@ google_cloud_cpp_common_srcs = [
     "iam_policy.cc",
     "internal/api_client_header.cc",
     "internal/backoff_policy.cc",
+    "internal/base64_transforms.cc",
     "internal/compiler_info.cc",
     "internal/credentials_impl.cc",
     "internal/filesystem.cc",
