+++ google/cloud/spanner/bytes_test.cc
@@ -29,10 +29,6 @@ namespace spanner {
 inline namespace SPANNER_CLIENT_NS {
 namespace {
 
-using ::google::cloud::testing_util::StatusIs;
-using ::testing::ContainsRegex;
-using ::testing::Not;
-
 TEST(Bytes, RoundTrip) {
   char c = std::numeric_limits<char>::min();
   std::string chars(1, c);
@@ -80,121 +76,6 @@ TEST(Bytes, RoundTrip) {
         EXPECT_EQ(data, bytes.get<std::string>());
       }
     }
-  }
-}
-
-TEST(Bytes, LongerRoundTrip) {
-  std::vector<std::pair<std::string, std::string>> test_cases = {
-      {"abcd", "YWJjZA=="},
-      {"abcde", "YWJjZGU="},
-      {"abcdef", "YWJjZGVm"},
-      {"abcdefg", "YWJjZGVmZw=="},
-      {"abcdefgh", "YWJjZGVmZ2g="},
-      {"abcdefghi", "YWJjZGVmZ2hp"},
-      {"abcdefghij", "YWJjZGVmZ2hpag=="},
-      {"abcdefghijk", "YWJjZGVmZ2hpams="},
-      {"abcdefghijkl", "YWJjZGVmZ2hpamts"},
-      {"abcdefghijklm", "YWJjZGVmZ2hpamtsbQ=="},
-      {"abcdefghijklmn", "YWJjZGVmZ2hpamtsbW4="},
-      {"abcdefghijklmno", "YWJjZGVmZ2hpamtsbW5v"},
-      {"abcdefghijklmnop", "YWJjZGVmZ2hpamtsbW5vcA=="},
-      {"abcdefghijklmnopq", "YWJjZGVmZ2hpamtsbW5vcHE="},
-      {"abcdefghijklmnopqr", "YWJjZGVmZ2hpamtsbW5vcHFy"},
-      {"abcdefghijklmnopqrs", "YWJjZGVmZ2hpamtsbW5vcHFycw=="},
-      {"abcdefghijklmnopqrst", "YWJjZGVmZ2hpamtsbW5vcHFyc3Q="},
-      {"abcdefghijklmnopqrstu", "YWJjZGVmZ2hpamtsbW5vcHFyc3R1"},
-      {"abcdefghijklmnopqrstuv", "YWJjZGVmZ2hpamtsbW5vcHFyc3R1dg=="},
-      {"abcdefghijklmnopqrstuvw", "YWJjZGVmZ2hpamtsbW5vcHFyc3R1dnc="},
-      {"abcdefghijklmnopqrstuvwx", "YWJjZGVmZ2hpamtsbW5vcHFyc3R1dnd4"},
-      {"abcdefghijklmnopqrstuvwxy", "YWJjZGVmZ2hpamtsbW5vcHFyc3R1dnd4eQ=="},
-      {"abcdefghijklmnopqrstuvwxyz", "YWJjZGVmZ2hpamtsbW5vcHFyc3R1dnd4eXo="},
-  };
-  for (auto const& test_case : test_cases) {
-    Bytes bytes(test_case.first);
-    EXPECT_EQ(test_case.second, spanner_internal::BytesToBase64(bytes));
-    EXPECT_EQ(test_case.first, bytes.get<std::string>());
-    auto decoded = spanner_internal::BytesFromBase64(test_case.second);
-    EXPECT_STATUS_OK(decoded) << test_case.first;
-    EXPECT_EQ(test_case.first, decoded->get<std::string>());
-    EXPECT_EQ(bytes, *decoded);
-  }
-}
-
-TEST(Bytes, RFC4648TestVectors) {
-  // https://tools.ietf.org/html/rfc4648#section-10
-  std::vector<std::pair<std::string, std::string>> test_cases = {
-      {"", ""},
-      {"f", "Zg=="},
-      {"fo", "Zm8="},
-      {"foo", "Zm9v"},
-      {"foob", "Zm9vYg=="},
-      {"fooba", "Zm9vYmE="},
-      {"foobar", "Zm9vYmFy"},
-  };
-  for (auto const& test_case : test_cases) {
-    Bytes bytes(test_case.first);
-    EXPECT_EQ(test_case.second, spanner_internal::BytesToBase64(bytes));
-    EXPECT_EQ(test_case.first, bytes.get<std::string>());
-    auto decoded = spanner_internal::BytesFromBase64(test_case.second);
-    EXPECT_STATUS_OK(decoded) << test_case.first;
-    EXPECT_EQ(test_case.first, decoded->get<std::string>());
-    EXPECT_EQ(bytes, *decoded);
-  }
-}
-
-TEST(Bytes, WikiExample) {
-  // https://en.wikipedia.org/wiki/Base64#Examples
-  std::string const plain =
-      "Man is distinguished, not only by his reason, but by this singular "
-      "passion from other animals, which is a lust of the mind, that by a "
-      "perseverance of delight in the continued and indefatigable generation "
-      "of knowledge, exceeds the short vehemence of any carnal pleasure.";
-  std::string const coded =
-      "TWFuIGlzIGRpc3Rpbmd1aXNoZWQsIG5vdCBvbmx5IGJ5IGhpcyByZWFzb24sIGJ1dCBieSB0"
-      "aGlzIHNpbmd1bGFyIHBhc3Npb24gZnJvbSBvdGhlciBhbmltYWxzLCB3aGljaCBpcyBhIGx1"
-      "c3Qgb2YgdGhlIG1pbmQsIHRoYXQgYnkgYSBwZXJzZXZlcmFuY2Ugb2YgZGVsaWdodCBpbiB0"
-      "aGUgY29udGludWVkIGFuZCBpbmRlZmF0aWdhYmxlIGdlbmVyYXRpb24gb2Yga25vd2xlZGdl"
-      "LCBleGNlZWRzIHRoZSBzaG9ydCB2ZWhlbWVuY2Ugb2YgYW55IGNhcm5hbCBwbGVhc3VyZS4"
-      "=";
-  Bytes bytes(plain);
-  EXPECT_EQ(coded, spanner_internal::BytesToBase64(bytes));
-  EXPECT_EQ(plain, bytes.get<std::string>());
-  auto decoded = spanner_internal::BytesFromBase64(coded);
-  EXPECT_STATUS_OK(decoded) << coded;
-  EXPECT_EQ(plain, decoded->get<std::string>());
-  EXPECT_EQ(bytes, *decoded);
-}
-
-TEST(Bytes, FromBase64Failures) {
-  // Bad lengths.
-  for (std::string const base64 : {"x", "xx", "xxx"}) {
-    auto decoded = spanner_internal::BytesFromBase64(base64);
-    EXPECT_THAT(decoded,
-                StatusIs(Not(StatusCode::kOk),
-                         ContainsRegex("Invalid base64.*at offset 0")));
-  }
-
-  for (std::string const base64 : {"xxxxx", "xxxxxx", "xxxxxxx"}) {
-    auto decoded = spanner_internal::BytesFromBase64(base64);
-    EXPECT_THAT(decoded,
-                StatusIs(Not(StatusCode::kOk),
-                         ContainsRegex("Invalid base64.*at offset 4")));
-  }
-
-  // Chars outside base64 alphabet.
-  for (std::string const base64 : {".xxx", "x.xx", "xx.x", "xxx.", "xx.="}) {
-    auto decoded = spanner_internal::BytesFromBase64(base64);
-    EXPECT_THAT(decoded,
-                StatusIs(Not(StatusCode::kOk),
-                         ContainsRegex("Invalid base64.*at offset 0")));
-  }
-
-  // Non-zero padding bits.
-  for (std::string const base64 : {"xx==", "xxx="}) {
-    auto decoded = spanner_internal::BytesFromBase64(base64);
-    EXPECT_THAT(decoded,
-                StatusIs(Not(StatusCode::kOk),
-                         ContainsRegex("Invalid base64.*at offset 0")));
   }
 }
